full auth is done but applied only to user routes
